@@ -27,27 +27,25 @@
     "author": "tmkn",
     "license": "MIT",
     "dependencies": {
-        "@emotion/react": "^11.8.1",
+        "@emotion/react": "^11.11.3",
         "@sentry/browser": "^6.18.1",
         "lodash.shuffle": "^4.2.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-router-dom": "^5.2.0",
-        "react-virtualized": "^9.22.3",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^6.21.2",
         "rxjs": "^7.5.4"
     },
     "devDependencies": {
         "@cypress/webpack-preprocessor": "^5.11.1",
         "@percy/cypress": "^3.1.1",
-        "@testing-library/jest-dom": "^5.16.2",
-        "@testing-library/react": "^12.1.3",
-        "@types/jest": "^27.4.1",
+        "@testing-library/jest-dom": "^6.2.0",
+        "@testing-library/react": "^14.1.2",
+        "@types/jest": "^29.5.11",
         "@types/lodash.shuffle": "^4.2.6",
         "@types/node": "^20.11.0",
-        "@types/react": "^17.0.38",
-        "@types/react-dom": "^17.0.11",
+        "@types/react": "^18.2.47",
+        "@types/react-dom": "^18.2.18",
         "@types/react-router-dom": "^5.3.3",
-        "@types/react-virtualized": "^9.21.20",
         "@webpack-cli/serve": "^1.6.1",
         "clean-webpack-plugin": "^4.0.0",
         "copy-webpack-plugin": "^12.0.1",
@@ -55,12 +53,13 @@
         "cypress": "^9.5.1",
         "express": "^4.17.3",
         "html-webpack-plugin": "^5.6.0",
-        "jest": "^27.5.1",
+        "jest": "^29.7.0",
+        "jest-environment-jsdom": "^29.7.0",
         "jest-fetch-mock": "^3.0.3",
         "prettier": "^3.2.2",
         "style-loader": "^3.3.4",
-        "ts-jest": "^27.1.3",
-        "ts-loader": "^9.2.6",
+        "ts-jest": "^29.1.1",
+        "ts-loader": "^9.5.1",
         "typescript": "^5.3.3",
         "webpack": "^5.89.0",
         "webpack-cli": "^5.1.4",
