@@ -1,1 +1,1 @@
-require('jest-fetch-mock').enableMocks()
+require("jest-fetch-mock").enableMocks();
