@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { jsx, css } from "@emotion/react";
-import React from "react";
+import React, { PropsWithChildren } from "react";
 
 import { primaryColor, mq } from "../../../css";
 
@@ -10,4 +10,6 @@ const highlightStyle = css({
     }
 });
 
-export const Highlight: React.FC = ({ children }) => <span css={highlightStyle}>{children}</span>;
+export const Highlight: React.FC<PropsWithChildren> = ({ children }) => (
+    <span css={highlightStyle}>{children}</span>
+);
